pyproject.toml
[project]
name = "paper-graph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests==2.31.0",
    "networkx==3.2.1",
    "matplotlib==3.8.2",
    "semanticscholar>=0.9.0",
    "python-dotenv==1.0.0",
    "rich==10.16.2",
    "dash==2.14.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
